tools/content-pack-builder/requirements.txt
# Content Pack Builder dependencies
# Python >= 3.10

sentence-transformers>=3.0.0
transformers>=4.40.0
torch>=2.0.0
faiss-cpu>=1.8.0
numpy>=1.24.0
